+++ src/app/(frontend)/page.tsx
@@ -1,4 +1,4 @@
-import BlogPostList from "@/src/components/blogpostlist/blog-post-list";
+import BlogPostList from "@/src/components/blog-post-list/blog-post-list";
 import { getAllPosts } from "@/src/sanity/lib/queries";
 
 const Home = () => {
